dtype -> 32

+++ deploy/deploy_real/act_to_dof.py
@@ -87,7 +87,8 @@ class ActToDof:
         if root_quat_wxyz is None:
             hands_command_b = hands_command_w
         else:
-            world_from_pelvis_quat = root_quat_wxyz
+            world_from_pelvis_quat = root_quat_wxyz.astype(np.float32)
+            hands_command_w = hands_command_w.astype(np.float32)
             hands_command_b = np.concatenate(
                 quat_rotate_inverse(world_from_pelvis_quat,
                                     hands_command_w[..., :3]),
